requirements.txt
fastapi~=0.120.0
uvicorn[standard]~=0.38.0
sqlmodel
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]

python-dotenv~=1.1.1
pydantic~=2.12.3